[#341] Use the same config storage for tests using mocks - ammended.
